+++ cineCablePro/src/pojo/annotations/custom/IngresosEgresosCierreCaja.java
@@ -11,21 +11,21 @@ public class IngresosEgresosCierreCaja implements Serializable {
 	private int idusuario;
 	private int idfpago;
 	private String nombre;
-	private float valpago;
+	private Double valpago;
 	private String ingresoegreso;
 	
 	public IngresosEgresosCierreCaja() {
 		
 	}
 	
-	public IngresosEgresosCierreCaja(int idusuario,int idfpago,String nombre,float valpago) {
+	public IngresosEgresosCierreCaja(int idusuario,int idfpago,String nombre,Double valpago) {
 		this.idusuario = idusuario;
 		this.idfpago = idfpago;
 		this.nombre = nombre;
 		this.valpago = valpago;
 	}
 	
-	public IngresosEgresosCierreCaja(int idusuario,int idfpago,String nombre,float valpago,String ingresoegreso) {
+	public IngresosEgresosCierreCaja(int idusuario,int idfpago,String nombre,Double valpago,String ingresoegreso) {
 		this.idusuario = idusuario;
 		this.idfpago = idfpago;
 		this.nombre = nombre;
@@ -57,11 +57,11 @@ public class IngresosEgresosCierreCaja implements Serializable {
 		this.nombre = nombre;
 	}
 
-	public float getValpago() {
+	public Double getValpago() {
 		return valpago;
 	}
 
-	public void setValpago(float valpago) {
+	public void setValpago(Double valpago) {
 		this.valpago = valpago;
 	}
